feat: change architecture to Atomic Design concept and add FormPage
