+++ src/App.tsx
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 import './App.css';
-import Home from './pages/Home';
-import About from './pages/About';
-import NotFound from './pages/NotFound';
+import Home from './pages/Home/Home';
+import About from './pages/About/About';
+import NotFound from './pages/404/NotFound';
 import MainLayout from './layouts/MainLayout';
+import FormPage from './pages/FormPage/FormPage';
 
 class App extends Component {
   render() {
@@ -15,7 +16,8 @@ class App extends Component {
             <Route path="/" element={<MainLayout />}>
               <Route index element={<Home />} />
               <Route path="about" element={<About />} />
-              <Route path="/404" element={<NotFound />} />
+              <Route path="form" element={<FormPage />} />
+              <Route path="404" element={<NotFound />} />
               <Route path="*" element={<Navigate to="/404" replace />} />
             </Route>
           </Routes>
