+++ src/components/atoms/icon/Icon.module.css
@@ -1,0 +1,9 @@
+.icon {
+  width: 24px;
+  height: 24px;
+  position: relative;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: 80%;
+  cursor: pointer;
+}
