+++ src/components/atoms/input/Input.module.css
@@ -1,0 +1,22 @@
+.input {
+  width: 100%;
+  padding: 10px;
+  font-size: 16px;
+  border: none;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px rgb(0 0 0 / 10%);
+  color: black;
+}
+
+.checkbox {
+  margin-right: 10px;
+  width: 20px;
+  height: 20px;
+  border-radius: 5px;
+  border: 1px solid black;
+  cursor: pointer;
+}
+
+.inputFile {
+  width: fit-content;
+}
