+++ src/components/atoms/typography/Span.module.css
@@ -1,0 +1,5 @@
+.span {
+  font-size: 14px;
+  font-weight: 600;
+  color: black;
+}
