+++ src/components/molecules/select/Select.module.css
@@ -1,0 +1,4 @@
+.select {
+  height: 30px;
+  align-self: center;
+}
