+++ src/components/organisms/Card/Card.module.css
@@ -1,0 +1,49 @@
+.imageContainer {
+  width: 100%;
+  height: 210px;
+  overflow: hidden;
+}
+
+.image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  background-color: aquamarine;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  font-size: 20px;
+}
+
+.content {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  height: 235px;
+}
+
+.CardTitle {
+  margin: 0 0 10px 0;
+  font-size: 1.2rem;
+}
+
+.cardDesc {
+  font-size: 1rem;
+  line-height: 1.5;
+  margin-bottom: 20px;
+}
+
+.cardButtons {
+  display: flex;
+  flex-wrap: wrap;
+  margin-bottom: 20px;
+  gap: 5px;
+}
+
+.cardFooter {
+  display: flex;
+  justify-content: space-between;
+  font-size: 0.9rem;
+  color: #808080;
+}
