+++ src/components/templates/Form/FormTemplate.module.css
@@ -1,0 +1,9 @@
+.form {
+  border-radius: 12px;
+  box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
+  padding: 25px;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 100px;
+  margin-top: 20px;
+}
