--- src/css/components/Card.module.css
@@ -1,60 +1,0 @@
-.card {
-  display: flex;
-  flex-direction: column;
-  background-color: white;
-  border-radius: 20px;
-  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.15);
-  width: 350px;
-  height: 485px;
-  overflow: hidden;
-}
-
-.imageContainer {
-  width: 100%;
-  height: 210px;
-  overflow: hidden;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  background-color: aquamarine;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  font-size: 20px;
-}
-
-.content {
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  height: 235px;
-}
-
-.CardTitle {
-  margin: 0 0 10px 0;
-  font-size: 1.2rem;
-}
-
-.cardDesc {
-  font-size: 1rem;
-  line-height: 1.5;
-  margin-bottom: 20px;
-}
-
-.cardButtons {
-  display: flex;
-  flex-wrap: wrap;
-  margin-bottom: 20px;
-  gap: 5px;
-}
-
-.cardFooter {
-  display: flex;
-  justify-content: space-between;
-  font-size: 0.9rem;
-  color: #808080;
-}
