--- src/css/components/CardButton.module.css
@@ -1,28 +1,0 @@
-.cardButton {
-  padding: 4px 12px;
-  background-color: #dbe4d1;
-  border-radius: 32px;
-  color: #fff;
-  border: none;
-  font-size: 14px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-}
-
-.cardButtonImg {
-  width: 24px;
-  height: 24px;
-  position: relative;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: 80%;
-  cursor: pointer;
-}
-
-.cardButtonCounter {
-  color: #626d7a;
-  font-size: 14px;
-  line-height: 16px;
-  font-weight: 500;
-}
