--- src/css/components/SearchBar.module.css
@@ -1,54 +1,0 @@
-.searchBar {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 80px;
-  width: 35%;
-  background-color: #f6f6f6;
-  border-radius: 50px;
-  padding: 0 20px;
-  margin: 30px auto;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.searchInput {
-  font-size: 18px;
-  color: #555;
-  background-color: transparent;
-  border: none;
-  outline: none;
-  margin-left: 20px;
-  flex: 1;
-}
-
-.searchIcon {
-  width: 24px;
-  height: 24px;
-  position: relative;
-  background: url(../../assets/search.svg);
-  background-repeat: no-repeat;
-  background-size: 100%;
-  cursor: pointer;
-  border: none;
-}
-
-.searchButton {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: darkslategray;
-  color: #fff;
-  font-size: 16px;
-  font-weight: bold;
-  text-transform: uppercase;
-  border: none;
-  border-radius: 50px;
-  width: 90px;
-  height: 40px;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-}
-
-.searchButton:hover {
-  background-color: #1d8181;
-}
