--- src/interfaces/ICardButton.ts
@@ -1,4 +1,0 @@
-export interface ICardButton {
-  cardButtonImg: string;
-  cardButtonCounter: number | string;
-}
