--- src/pages/About.tsx
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-
-class About extends Component {
-  render() {
-    return (
-      <>
-        <header>
-          <h1>About</h1>
-        </header>
-        <p className="empty__page-content">About page content</p>
-      </>
-    );
-  }
-}
-
-export default About;
